+++ src/main/java/com/influy/domain/item/domain/ItemStatus.java
@@ -1,4 +1,4 @@
-package com.influy.domain.item;
+package com.influy.domain.item.domain;
 
 public enum ItemStatus {
     DEFAULT, SOLDOUT, ONE,TWO, THREE, FOUR, FIVE, SIX, SEVEN
