+++ src/main/java/com/influy/domain/profileLink/domain/ProfileLinkType.java
@@ -1,4 +1,4 @@
-package com.influy.domain.profileLink;
+package com.influy.domain.profileLink.domain;
 
 public enum ProfileLinkType {
     YOUTUBE,INSTAGRAM,TIKTOK,EMAIL,ETC
