+++ src/main/java/com/influy/domain/seller/domain/Seller.java
@@ -1,6 +1,5 @@
-package com.influy.domain.seller;
+package com.influy.domain.seller.domain;
 
-import com.influy.domain.item.Item;
 import com.influy.global.common.BaseEntity;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
@@ -8,8 +7,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.util.List;
 
 @Entity
 @Builder
