+++ src/main/java/com/influy/domain/user/domain/User.java
@@ -1,4 +1,4 @@
-package com.influy.domain.user;
+package com.influy.domain.user.domain;
 
 import com.influy.global.common.BaseEntity;
 import jakarta.persistence.Entity;
